requirements.txt
django
djangorestframework
python-decouple
psycopg2-binary
djangorestframework_simplejwt
celery[redis]